requirements.txt
fastapi
fastapi[standard]
schwabdev
redis
python-dotenv
polygon-api-client
cdp-sdk
coinbase-advanced-py